+++ CHANGELOG.md
@@ -7,6 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.1.14] - 2025-05-30
+
+### Fixed
+
+- Data cleansing now automatically removes leading and trailing whitespace from string columns
+- Fixed chat endpoint when DATAROBOT_ENDPOINT has a trailing slash
+- Fixed the raw data preview for SAP (react)
+
+### Added
+
+- Ability to download a specific chat history (including charts) in the React version.
+
+### Changed
+
+- React-based Frontend as the default for the Application
+- Improved error handling when prompting (react)
+- Change react frontend `deploy` app to use AF fastapi template
+- Change react frontend to use AF react template, removes `frontend_react` in favor of `app_frontend`
+
 ## [0.1.13] - 2025-05-06
 
 ### Fixed